% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_transformation.R
\name{get_transformation}
\alias{get_transformation}
\title{Apply \code{bestNormalize} Transformation to Variables in a Dataset and Update Dataset Accordingly}
\usage{
get_transformation(data, var_list, seed = 100)
}
\arguments{
\item{data}{A dataframe to which the \code{bestNormalize} function should be applied.}

\item{var_list}{A character vector of variable names.}

\item{seed}{A single value, interpreted as an integer, to set the random seed for reproducibility.}
}
\value{
A list with two elements: \code{data} containing the updated dataframe and \code{var_list} containing
the updated variable list.
}
\description{
The \code{transform_and_update} function takes a dataset and a vector of variable names,
applies the \code{bestNormalize} function to each variable, creates transformed versions of the
variables when necessary, and updates the dataset and the variable list accordingly.
It returns a list containing the updated dataset and the updated variable list.
}
\examples{
\dontrun{
# Generate some data
df <- data.frame(
  daily_sat_aut = rnorm(100),
  daily_fru_aut = rnorm(100),
  daily_sat_rel = rnorm(100),
  daily_fru_rel = rnorm(100)
)

# Apply the `bestNormalize` function and update dataset
result <- transform_and_update(df, var_list = c("daily_sat_aut", "daily_fru_aut", "daily_sat_rel", "daily_fru_rel"))
updated_df <- result$data
updated_var_list <- result$var_list
}

}
