% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_and_load_git_packages.R
\name{install_and_load_git_packages}
\alias{install_and_load_git_packages}
\title{Install and Load GitHub Packages}
\usage{
install_and_load_git_packages(git_packages)
}
\arguments{
\item{git_packages}{A vector of GitHub repositories corresponding to the packages to be installed
and loaded. Each repository is specified as a string in the format "username/repository".}
}
\value{
No return value. Called for side effects.
}
\description{
This function installs and loads R packages from GitHub repositories.
It checks if the specified packages are already installed. If a package is not
installed, it is downloaded, installed, and then loaded into the R environment.
}
\examples{
\dontrun{
install_and_load_git_packages(c("CharlesEtienneLavoie/CEL", 'rempsyc/rempsyc'))
}
}
\references{
Remember to add reference here.
}
