% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_and_load_packages.R
\name{install_and_load_packages}
\alias{install_and_load_packages}
\title{Install and Load R Packages}
\usage{
install_and_load_packages(packages)
}
\arguments{
\item{packages}{A character vector of package names to be installed and loaded.}
}
\value{
No return value. Called for side effects.
}
\description{
This function installs and loads R packages. It checks if the specified packages
are already installed. If a package is not installed, it is downloaded from CRAN,
installed, and then loaded into the R environment.
}
\examples{
\dontrun{
install_and_load_packages(c("ggplot2", "dplyr"))
}
}
\references{
Remember to add reference here.
}
