% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list2string.R
\name{list2string}
\alias{list2string}
\title{Convert List to String}
\usage{
list2string(var_list)
}
\arguments{
\item{var_list}{A list or vector of items to be converted to a string.}
}
\value{
Prints out the string representation of the input list.
However, the function itself does not have a return value; it is called for its side effects.
}
\description{
This function takes a list of variables as input and converts them to a string representation,
enclosed within \code{c()}. This could be useful when dealing with the formatted printing of a vector/list
in a user-friendly manner.
}
\examples{
\dontrun{
list2string(c("var1", "var2", "var3"))
}
}
\references{
Remember to add reference here
}
