% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_qualtrics_files.R
\name{process_qualtrics_files}
\alias{process_qualtrics_files}
\title{Process Qualtrics Data Files}
\usage{
process_qualtrics_files(data_files, rename_files = FALSE, new_names = NULL)
}
\arguments{
\item{data_files}{A character vector of file paths to the CSV files to be
processed.}

\item{rename_files}{Logical. If TRUE, the data and labels for each file will
be saved under a new name based on the original file name. Default is FALSE.}

\item{new_names}{A character vector of new names for the files. This should
be the same length as data_files. If NULL and rename_files is TRUE, new names
will be generated by removing the timestamp and anything before it from the
original file name. Default is NULL.}
}
\value{
This function does not return anything. It writes processed data and
labels to CSV files and assigns the resulting data frames to new variables in
the global environment.
}
\description{
This function reads in and processes a series of CSV files
exported from Qualtrics. It assumes the first two rows of each file contain
metadata and not actual data.
}
\examples{
\dontrun{
process_qualtrics_files(c("file1.csv", "file2.csv"), rename_files = TRUE, new_names = c("survey1", "survey2"))
}
}
\references{
Remember to add reference here
}
