% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_env_objects.R
\name{remove_env_objects}
\alias{remove_env_objects}
\title{Remove Specific Objects from Global Environment}
\usage{
remove_env_objects(remove_objects = NULL)
}
\arguments{
\item{remove_objects}{A character vector specifying the names of objects to
be removed. If NULL (default), no objects are removed.}
}
\value{
This function does not return anything. It removes the specified
objects from the global environment.
}
\description{
This function removes specified objects from the global
environment in R.
}
\examples{
\dontrun{
# create some objects
x <- 1:10
y <- letters
z <- rnorm(100)

# remove 'x' and 'y'
remove_env_objects(c("x", "y"))
}
}
\references{
Remember to add reference here
}
