% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_additional_demographics.R
\name{report_additional_demographics}
\alias{report_additional_demographics}
\title{Report Descriptive Statistics for Demographics}
\usage{
report_additional_demographics(
  data,
  income = NULL,
  empl_status = NULL,
  alcohol_units = NULL
)
}
\arguments{
\item{data}{A dataframe that contains the demographic variables.}

\item{income}{A string that represents the name of the income variable in the data (default: NULL).}

\item{empl_status}{A string that represents the name of the employment status variable in the data (default: NULL).}

\item{alcohol_units}{A string that represents the name of the weekly alcohol units variable in the data (default: NULL).}
}
\value{
A string containing the descriptive statistics for each demographic variable provided.
}
\description{
The \code{report_additional_demographics} function generates a summary
report of descriptive statistics (counts and percentages) for demographic variables
such as income, employment status, and weekly alcohol units.
}
\details{
The function allows each demographic to be optional.
If a demographic parameter is set to null (default), the function will not perform any
operations related to that demographic. If a valid demographic variable name
is passed as a parameter, the function will calculate the number of instances and
the corresponding percentage for each category in the given demographic variable.
}
\examples{
\dontrun{
data <- data.frame(
 demo_income_c = sample(c("$100,001 - $125,000", "$125,001 - $150,000"), 100, replace = TRUE),
 demo_empl_status_c = sample(c("Full-time", "Part-time"), 100, replace = TRUE),
 demo_alcohol_units_c = sample(c("1-4", "5-9"), 100, replace = TRUE)
)

report_additional_demographics(
  data = data,
  income = "demo_income_c",
  empl_status = "demo_empl_status_c",
  alcohol_units = "demo_alcohol_units_c"
)
}

}
