% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_means.R
\name{get_means}
\alias{get_means}
\title{Compute Mean Scores for Scales with Optional Reverse Scoring and Clustering}
\usage{
get_means(
  data,
  scale_list,
  reverse_score = TRUE,
  rename_reversed_items = FALSE,
  order = 1,
  cluster = NULL
)
}
\arguments{
\item{data}{A data frame containing the scales and items.}

\item{scale_list}{A character vector with the base names of the scales to be computed.}

\item{reverse_score}{A logical indicating whether to apply reverse scoring. Default is TRUE.}

\item{rename_reversed_items}{A logical indicating whether to rename reverse
scored items by replacing "reverse" with "reversed". Default is FALSE.}

\item{order}{An integer indicating the order of the scale. Use 1 for first order scales
and 2 for second order scales. Default is 1.}

\item{cluster}{A character representing the name of the column to be used for clustering.
If provided, the function will compute cluster-level means for each new variable. Default is NULL.}
}
\value{
A data frame with the computed mean scale scores added as new columns.
If a cluster variable was provided, additional cluster-level mean variables are also added.
}
\description{
This function computes mean scores for specified scales from a
given list. The scales are expected to be columns in the data frame starting
with a base name in the scale list. If a scale item needs to be reverse
scored (identified by "reverse" or "Reverse" in the item name), it is done
so before the mean is computed. An option allows renaming of reverse scored items.
If a cluster variable is provided, the function also computes cluster-level means for
each newly computed variable.
}
