% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_and_load_packages.R
\name{install_and_load_packages}
\alias{install_and_load_packages}
\title{Compute Power Analysis For SEM Model Using Monte Carlo Simulation}
\usage{
install_and_load_packages(packages)
}
\arguments{
\item{model.population}{The lavaan model with population estimate values
specified.}

\item{model}{The lavaan model that will be tested in each simulation}

\item{ksim}{How many simulations the function should perform}

\item{nobs}{How many observations to generate in each simulation}
}
\value{
A table.
}
\description{
Compute Power Analysis For SEM Model Using Monte Carlo
Simulation
}
\details{
The function uses the simulate_data function from the lavaan
package to perform a monte carlo simulation. The mean, std error, z value,
p value and confidence intervals are then computed and reported in a table
for each parameter.
}
\examples{
library(lavaan)
modpop <- '
M ~ 0.40*X
Y ~ 0.30*M
'
mod <- '
M ~ X
Y ~ M
'

simulate_power(modpop, mod)
}
\references{
Remember to add reference here
}
